requirements.txt
numpy
optuna
wandb
tensorboard
pyyaml
plotly
transformers==4.1.1
petastorm==0.9.2
scikit-learn==0.23.2
torch==1.8.1
fastai==1.0.61
cufflinks==0.17.3
